move tests from test to relative __tests__ directory
