+++ src/merkle/__tests__/merkle-layers.test.ts
@@ -1,8 +1,8 @@
 import { assert } from 'chai';
 import { describe, it } from 'mocha';
 
-import { MergeFunction, Node, CompareFunction } from '../../src/merkle/merkle';
-import { MerkleTree } from '../../src/merkle/merkle-tree';
+import { MergeFunction, Node, CompareFunction } from '../merkle';
+import { MerkleTree } from '../merkle-tree';
 
 class StringConcat implements MergeFunction<string> {
   async merge(n1: Node<string>, n2: Node<string>): Promise<Node<string>> {
